Moved the parameter checking jwt middlware into paramcheck.js as a middlware.That way we can export it to the plaid router ezpz
